genrate files
-copy files function created that given a source path  and destination
path recursively copies files from source to destination.
-generate page function made to create an html page from markdown.Opens
and read markdwon file which is converted to html using
markdown_to_html.A template file containing everything needed to render
a html page in the browser is read and pre placed areas are replaced for
desired values.
-generate_page_recursively function created to generate html pages from
a whole directory using the functionality of generate_page.Function
checks for any md files creating html files for each and complimentary
directories in a public folder.